Add files via upload
Hey there, future explorers! Have you ever looked up at the night sky and wondered what secrets it holds? What if there are worlds out there waiting to be discovered, each with its own mysteries and adventures? 


That’s why we’re super excited to introduce you to Sky Breakers, our amazing web game! It’s designed just for you to dive into the wonders of the universe in a fun and interactive way. 


But before we blast off, let’s meet the awesome team behind Sky Breakers! We’re a group of passionate space enthusiasts and game developers who believe learning should be an adventure. Together, we’ve created a game that makes exploring exoplanets exciting and easy to understand. 


So, are you ready to join us on this incredible journey? Let’s unlock the secrets of the sky together in Sky Breakers!